toml-stream: don't make unnecessary Promises
Emitting each object as TOML is actually a completely synchronous
operation as far as the TOMLStream is concerned. There's probably a
better way to handle backpressure, seeing as a large object is going to
generate a flurry of pushes, but for now, this is far simpler than the
old code, even if it does use ES6 inheritance.